Eliminate several command queue / Turbo Queue related struct fields.
Several structure fields that were only needed to support older SiS
chips were removed.  The single remaining use of RealVideoRam was
replaced with 'pscrn->videoRam'.

The code removed in xgi_vga.c was removed because the corresponding
code in xf86-video-sis is only used on some SiS 3xx chips.  The XG20 /
XG40 support in that driver does not do this.