Clean up some of the vertical retrace madness.
There were (are) several routines floating around in the code that all do
basically the same thing...wait for the vertical refresh.  I've cleaned up
some of it, primarilly moving prototype to sane places and deleting dead
code.

Threre are still two similar but subtly different routines that wait for the
vertical refresh.  At some point we'll need to evaluate whether or not both
are really necessary.