Enable DRI when XGIfb is not available.
Previously DRI was only enabled when XGIfb was available.  This was
done because XGIfb was the only place that initialized the framebuffer
memory manager.  However, there is an ioctl for this purpose.  Code
was added to xgi_dri.c to call this ioctl.

This matches the way the SiS driver works.