Change default configuration for maxxfbmem
Previously maxxfbmem had to be configured by the user via xorg.conf in
order for DRI to work.  maxxfbmem is now set to the required size for
the virtual display.  This allows DRI to work without additional user
intervention.

In addition, code was added to detect the case there no additional
framebuffer memory is available.  In this case, DRI is disabled.  This
prevents mysterious lock-ups when trying to use DRI without video
memory for back buffers, etc.